refactor(app): split main app into modules
